refactor(kubernautagent): split parser.go into focused files
Extracts JSON extraction helpers (extractJSON, extractBalancedJSON,
scanToBalancedClose) into parser_json_extract.go, the nested LLM response
types and double-serialization/coercion helpers (llmResponse, llmRCA,
unwrapDoubleSerializedJSON, coerceKnownFields, etc.) into
parser_llm_types.go, and the format-parsing/merge functions
(parseLLMFormat, parseSectionHeaders, mergeNested*, applyFlatFields) into
parser_format.go. Reduces parser.go from 872 to 153 lines (ResultParser +
Parse + ApplyInvestigationOutcome only), no behavioral changes. goimports
reordered an import in parser_test.go.
