+++ internal/kubernautagent/parser/parser.go
@@ -18,9 +18,6 @@ package parser
 
 import (
 	"encoding/json"
-	"fmt"
-	"strconv"
-	"strings"
 
 	"github.com/go-logr/logr"
 
@@ -90,725 +87,9 @@ func (p *ResultParser) Parse(content string) (*katypes.InvestigationResult, erro
 	return nil, &ErrNoRecognizedFields{Raw: jsonStr}
 }
 
-// extractJSON finds JSON content using a priority chain:
-// 1. Fenced ```json ... ``` code blocks
-// 2. Fenced ``` ... ``` code blocks containing JSON
-// 3. Raw string starting with {
-// 4. Balanced brace extraction (GAP-003: handles JSON embedded in prose)
-func extractJSON(content string) string {
-	if idx := strings.Index(content, "```json"); idx != -1 {
-		start := idx + len("```json")
-		end := strings.Index(content[start:], "```")
-		if end != -1 {
-			return strings.TrimSpace(content[start : start+end])
-		}
-	}
-	if idx := strings.Index(content, "```"); idx != -1 {
-		start := idx + len("```")
-		end := strings.Index(content[start:], "```")
-		if end != -1 {
-			candidate := strings.TrimSpace(content[start : start+end])
-			if len(candidate) > 0 && candidate[0] == '{' {
-				return candidate
-			}
-		}
-	}
-	trimmed := strings.TrimSpace(content)
-	if len(trimmed) > 0 && trimmed[0] == '{' {
-		return trimmed
-	}
-	if len(trimmed) > 0 && trimmed[0] == '[' {
-		return unwrapSingleElementArray(trimmed)
-	}
-	return extractBalancedJSON(content)
-}
-
-// unwrapSingleElementArray handles LLMs that wrap their response in a JSON
-// array (e.g. `[{"rca_summary":...}]`). Single-element arrays are unwrapped
-// to the inner object; multi-element arrays are rejected as ambiguous.
-func unwrapSingleElementArray(s string) string {
-	var arr []json.RawMessage
-	if err := json.Unmarshal([]byte(s), &arr); err != nil {
-		return ""
-	}
-	if len(arr) != 1 {
-		return ""
-	}
-	elem := strings.TrimSpace(string(arr[0]))
-	if len(elem) > 0 && elem[0] == '{' {
-		return elem
-	}
-	return ""
-}
-
-// extractBalancedJSON finds the first complete JSON object in content
-// by counting balanced braces. Handles JSON embedded in prose text,
-// mirroring KA's json_utils.py balanced extraction.
-//
-// M3-fix: skip `{` chars that are likely prose (not followed by `"`, `\n`, or
-// another `{`). Real JSON objects start with `{"` or `{\n`.
-func extractBalancedJSON(content string) string {
-	pos := 0
-	for pos < len(content) {
-		start := strings.IndexByte(content[pos:], '{')
-		if start == -1 {
-			return ""
-		}
-		start += pos
-
-		if !looksLikeJSONObjectStart(content, start) {
-			pos = start + 1
-			continue
-		}
-
-		if end := scanToBalancedClose(content, start); end != -1 {
-			return content[start:end]
-		}
-		pos = start + 1
-	}
-	return ""
-}
-
-// looksLikeJSONObjectStart reports whether the '{' at content[start] is
-// plausibly the start of a JSON object, rather than a stray brace embedded
-// in prose (e.g. "the {foo} placeholder"). A real JSON object's opening
-// brace is followed by a quote, whitespace, or another brace.
-func looksLikeJSONObjectStart(content string, start int) bool {
-	if start+1 >= len(content) {
-		return true
-	}
-	switch content[start+1] {
-	case '"', '\n', '\r', ' ', '\t', '{', '}':
-		return true
-	default:
-		return false
-	}
-}
-
-// scanToBalancedClose scans content starting at the opening brace index
-// start, tracking string-literal state and escape sequences, and returns the
-// index one past the matching closing brace (i.e. a valid content[start:end]
-// slice bound). Returns -1 if the braces never balance before EOF.
-func scanToBalancedClose(content string, start int) int {
-	depth := 0
-	inString := false
-	escaped := false
-
-	for i := start; i < len(content); i++ {
-		ch := content[i]
-		if escaped {
-			escaped = false
-			continue
-		}
-		if ch == '\\' && inString {
-			escaped = true
-			continue
-		}
-		if ch == '"' {
-			inString = !inString
-			continue
-		}
-		if inString {
-			continue
-		}
-		switch ch {
-		case '{':
-			depth++
-		case '}':
-			depth--
-			if depth == 0 {
-				return i + 1
-			}
-		}
-	}
-	return -1
-}
-
 const notActionableWarning = "Alert not actionable — no remediation warranted"
 const problemResolvedWarning = "Problem self-resolved"
 const confidenceFloor = 0.8
-
-// llmResponse is the nested JSON structure that LLMs typically produce.
-// Fields here must stay in sync with the JSON schema in schema.go and the
-// structured output prompt template in incident_investigation.tmpl.
-type llmResponse struct {
-	RCA                  *llmRCA                `json:"root_cause_analysis"`
-	RCAAlt               *llmRCA                `json:"rootCauseAnalysis,omitempty"`
-	Workflow             *llmWorkflow           `json:"selected_workflow"`
-	AlternativeWorkflows []llmAlternative       `json:"alternative_workflows,omitempty"`
-	Severity             string                 `json:"severity,omitempty"`
-	Confidence           float64                `json:"confidence,omitempty"`
-	Actionable           *bool                  `json:"actionable,omitempty"`
-	InvestigationOutcome string                 `json:"investigation_outcome,omitempty"`
-	DetectedLabels       map[string]interface{} `json:"detected_labels,omitempty"`
-}
-
-// resolvedRCA returns the RCA from either snake_case or camelCase key.
-// #746: LLMs sometimes use camelCase rootCauseAnalysis instead of snake_case.
-func (r *llmResponse) resolvedRCA() *llmRCA {
-	if r.RCA != nil {
-		return r.RCA
-	}
-	return r.RCAAlt
-}
-
-type llmAlternative struct {
-	WorkflowID string                 `json:"workflow_id"`
-	Confidence float64                `json:"confidence"`
-	Rationale  string                 `json:"rationale"`
-	Parameters map[string]interface{} `json:"parameters,omitempty"`
-}
-
-type llmRCA struct {
-	Summary               string                      `json:"summary"`
-	Severity              string                      `json:"severity,omitempty"`
-	SignalName            string                      `json:"signal_name,omitempty"`
-	ContributingFactors   []string                    `json:"contributing_factors,omitempty"`
-	RemediationTarget     *llmRemTarget               `json:"remediation_target,omitempty"`
-	RemediationTargetAlt  *llmRemTarget               `json:"remediationTarget,omitempty"`
-	InvestigationAnalysis string                      `json:"investigation_analysis,omitempty"`
-	CausalChain           []string                    `json:"causal_chain,omitempty"`
-	DueDiligence          *katypes.DueDiligenceReview `json:"due_diligence,omitempty"`
-}
-
-func (r *llmRCA) resolvedTarget() *llmRemTarget {
-	if r.RemediationTarget != nil {
-		return r.RemediationTarget
-	}
-	return r.RemediationTargetAlt
-}
-
-type llmRemTarget struct {
-	Kind       string `json:"kind"`
-	Name       string `json:"name"`
-	Namespace  string `json:"namespace"`
-	APIVersion string `json:"api_version,omitempty"`
-}
-
-type llmWorkflow struct {
-	WorkflowID      string                 `json:"workflow_id"`
-	ExecutionBundle string                 `json:"execution_bundle,omitempty"`
-	Confidence      float64                `json:"confidence"`
-	Rationale       string                 `json:"rationale,omitempty"`
-	Parameters      map[string]interface{} `json:"parameters,omitempty"`
-	ExecutionEngine string                 `json:"execution_engine,omitempty"`
-}
-
-// unwrapDoubleSerializedJSON detects when an LLM has double-serialized a
-// structured field (e.g. root_cause_analysis: "{\"summary\":...}" instead of
-// root_cause_analysis: {"summary":...}) and returns corrected JSON. This
-// defends against LLMs that json.Marshal their tool call arguments before
-// embedding them (Issue #795).
-func unwrapDoubleSerializedJSON(rawJSON string, logger logr.Logger) string {
-	var raw map[string]json.RawMessage
-	if err := json.Unmarshal([]byte(rawJSON), &raw); err != nil {
-		return rawJSON
-	}
-
-	changed := false
-	for _, key := range []string{"root_cause_analysis", "rootCauseAnalysis", "selected_workflow"} {
-		val, ok := raw[key]
-		if !ok {
-			continue
-		}
-		var s string
-		if json.Unmarshal(val, &s) != nil {
-			continue
-		}
-		s = strings.TrimSpace(s)
-		if len(s) == 0 || s[0] != '{' {
-			continue
-		}
-		if t := extractBalancedJSON(s); t != "" && json.Valid([]byte(t)) {
-			if len(t) != len(s) {
-				logger.V(1).Info("unwrapDoubleSerializedJSON: stripped trailing content",
-					"key", key,
-					"original_len", len(s),
-					"extracted_len", len(t))
-			}
-			raw[key] = json.RawMessage(t)
-			changed = true
-		}
-	}
-
-	if !changed {
-		return rawJSON
-	}
-	fixed, err := json.Marshal(raw)
-	if err != nil {
-		return rawJSON
-	}
-	return string(fixed)
-}
-
-// coerceKnownFields fixes type drift from LLMs that return numeric or boolean
-// fields as quoted strings (e.g. "confidence":"0.92" or "actionable":"false").
-// It operates on raw JSON bytes, unquoting known fields before typed unmarshal.
-func coerceKnownFields(rawJSON string) string {
-	var raw map[string]json.RawMessage
-	if err := json.Unmarshal([]byte(rawJSON), &raw); err != nil {
-		return rawJSON
-	}
-
-	changed := coerceStringifiedNumberField(raw, "confidence")
-	changed = coerceStringifiedBoolField(raw, "actionable") || changed
-	changed = coerceNestedWorkflowConfidence(raw) || changed
-
-	if !changed {
-		return rawJSON
-	}
-	fixed, err := json.Marshal(raw)
-	if err != nil {
-		return rawJSON
-	}
-	return string(fixed)
-}
-
-// coerceStringifiedNumberField unquotes raw[key] in place when it holds a
-// quoted numeric string (e.g. "0.92"), or deletes the key when the quoted
-// value doesn't parse as a number. Reports whether raw was touched.
-func coerceStringifiedNumberField(raw map[string]json.RawMessage, key string) bool {
-	val, ok := raw[key]
-	if !ok {
-		return false
-	}
-	var s string
-	if json.Unmarshal(val, &s) != nil {
-		return false
-	}
-	s = strings.TrimSpace(s)
-	if _, err := strconv.ParseFloat(s, 64); err == nil {
-		raw[key] = json.RawMessage(s)
-	} else {
-		delete(raw, key)
-	}
-	return true
-}
-
-// coerceStringifiedBoolField unquotes raw[key] in place when it holds a
-// quoted "true"/"false" string, or deletes the key otherwise. Reports
-// whether raw was touched.
-func coerceStringifiedBoolField(raw map[string]json.RawMessage, key string) bool {
-	val, ok := raw[key]
-	if !ok {
-		return false
-	}
-	var s string
-	if json.Unmarshal(val, &s) != nil {
-		return false
-	}
-	s = strings.TrimSpace(s)
-	if s == "true" || s == "false" {
-		raw[key] = json.RawMessage(s)
-	} else {
-		delete(raw, key)
-	}
-	return true
-}
-
-// coerceNestedWorkflowConfidence applies the same stringified-number fix to
-// selected_workflow.confidence, re-marshaling the nested object back into
-// raw when changed. Reports whether raw was touched.
-func coerceNestedWorkflowConfidence(raw map[string]json.RawMessage) bool {
-	nested, ok := raw["selected_workflow"]
-	if !ok {
-		return false
-	}
-	var wf map[string]json.RawMessage
-	if json.Unmarshal(nested, &wf) != nil {
-		return false
-	}
-	if !coerceStringifiedNumberFieldIfQuoted(wf, "confidence") {
-		return false
-	}
-	fixed, err := json.Marshal(wf)
-	if err != nil {
-		return false
-	}
-	raw["selected_workflow"] = json.RawMessage(fixed)
-	return true
-}
-
-// coerceStringifiedNumberFieldIfQuoted mirrors coerceStringifiedNumberField's
-// unquote behavior but only reports a change when the field was present,
-// quoted, and successfully parsed as a number — it leaves unparsable values
-// in place rather than deleting them (matching the original nested-workflow
-// coercion, which only ever touched confidence on a successful parse).
-func coerceStringifiedNumberFieldIfQuoted(raw map[string]json.RawMessage, key string) bool {
-	val, ok := raw[key]
-	if !ok {
-		return false
-	}
-	var s string
-	if json.Unmarshal(val, &s) != nil {
-		return false
-	}
-	s = strings.TrimSpace(s)
-	if _, err := strconv.ParseFloat(s, 64); err != nil {
-		return false
-	}
-	raw[key] = json.RawMessage(s)
-	return true
-}
-
-// flatLLMFields captures top-level fields that may appear alongside the flat
-// InvestigationResult format (rca_summary, workflow_id, confidence, etc.).
-type flatLLMFields struct {
-	Severity             string `json:"severity,omitempty"`
-	Actionable           *bool  `json:"actionable,omitempty"`
-	InvestigationOutcome string `json:"investigation_outcome,omitempty"`
-}
-
-// applyLLMRCA copies the resolved RCA fields (whichever of snake_case /
-// camelCase the LLM used) onto result, logging when the remediation target
-// is missing its required api_version.
-func applyLLMRCA(result *katypes.InvestigationResult, rca *llmRCA, logger logr.Logger) {
-	if rca == nil {
-		return
-	}
-	result.RCASummary = rca.Summary
-	result.Severity = rca.Severity
-	result.SignalName = rca.SignalName
-	result.ContributingFactors = rca.ContributingFactors
-	result.InvestigationAnalysis = rca.InvestigationAnalysis
-	result.CausalChain = rca.CausalChain
-	result.DueDiligence = rca.DueDiligence
-
-	t := rca.resolvedTarget()
-	if t == nil {
-		return
-	}
-	if t.APIVersion == "" && t.Kind != "" {
-		logger.Info("LLM omitted required api_version for remediation_target, resolution will use heuristic fallback",
-			"kind", t.Kind, "name", t.Name, "namespace", t.Namespace)
-	}
-	result.RemediationTarget = katypes.RemediationTarget{
-		Kind:       t.Kind,
-		Name:       t.Name,
-		Namespace:  t.Namespace,
-		APIVersion: t.APIVersion,
-	}
-}
-
-// applyLLMWorkflow copies the selected-workflow fields onto result, merging
-// (rather than replacing) any parameters already present.
-func applyLLMWorkflow(result *katypes.InvestigationResult, wf *llmWorkflow) {
-	if wf == nil {
-		return
-	}
-	result.WorkflowID = wf.WorkflowID
-	result.ExecutionBundle = wf.ExecutionBundle
-	result.Confidence = wf.Confidence
-	result.Reason = wf.Rationale
-	result.WorkflowRationale = wf.Rationale
-	result.ExecutionEngine = wf.ExecutionEngine
-	if wf.Parameters == nil {
-		return
-	}
-	if result.Parameters == nil {
-		result.Parameters = make(map[string]interface{})
-	}
-	for k, v := range wf.Parameters {
-		result.Parameters[k] = v
-	}
-}
-
-// parseLLMFormat parses the nested LLM response format and converts
-// it to a flat InvestigationResult.
-func parseLLMFormat(jsonStr string, logger logr.Logger) (*katypes.InvestigationResult, error) {
-	fixed := unwrapDoubleSerializedJSON(jsonStr, logger)
-	fixed = coerceKnownFields(fixed)
-
-	var resp llmResponse
-	if err := json.Unmarshal([]byte(fixed), &resp); err != nil {
-		return nil, fmt.Errorf("parsing LLM JSON response: %w", err)
-	}
-
-	result := &katypes.InvestigationResult{}
-	applyLLMRCA(result, resp.resolvedRCA(), logger)
-
-	// Top-level severity takes precedence over nested (allows Mock LLM to set both)
-	if resp.Severity != "" {
-		result.Severity = resp.Severity
-	}
-	// Top-level confidence serves as fallback when selected_workflow is absent
-	// (e.g., not-actionable outcomes where the LLM still provides a confidence score).
-	if resp.Confidence > 0 {
-		result.Confidence = resp.Confidence
-	}
-	applyLLMWorkflow(result, resp.Workflow)
-
-	for _, alt := range resp.AlternativeWorkflows {
-		result.AlternativeWorkflows = append(result.AlternativeWorkflows, katypes.AlternativeWorkflow{
-			WorkflowID: alt.WorkflowID,
-			Confidence: alt.Confidence,
-			Rationale:  alt.Rationale,
-			Parameters: alt.Parameters,
-		})
-	}
-
-	applyFlatFields(result, flatLLMFields{
-		Severity:             resp.Severity,
-		Actionable:           resp.Actionable,
-		InvestigationOutcome: resp.InvestigationOutcome,
-	})
-
-	if len(resp.DetectedLabels) > 0 {
-		result.DetectedLabels = resp.DetectedLabels
-	}
-
-	// #746: Relax guard to accept responses where the LLM provided at least one
-	// recognizable signal (confidence, RCA, or workflow). HAPI v1.2.1 has no such
-	// guard — all parsed JSON flows through outcome routing. We require confidence > 0
-	// as a minimum to reject truly garbage JSON (e.g., {"foo": "bar"}).
-	hasContent := result.RCASummary != "" || result.WorkflowID != "" || resp.Confidence > 0
-	if !hasContent {
-		return nil, &ErrNoRecognizedFields{}
-	}
-
-	// #795 Fix D: Confidence alone without summary or workflow is a partial parse
-	// caused by json.Unmarshal silently skipping type-mismatched fields or the inner
-	// RCA object missing the required "summary" field. Reject so the investigator's
-	// retryRCASubmit can request the missing data. See #795 preflight audit.
-	if resp.Confidence > 0 && result.RCASummary == "" && result.WorkflowID == "" {
-		return nil, &ErrNoRecognizedFields{}
-	}
-
-	return result, nil
-}
-
-// parseSectionHeaders handles the "# header\nvalue" format produced by LLMs when
-// structured output (output_config) is unavailable (e.g. Vertex AI).
-// The prompt instructs:
-//
-//	# root_cause_analysis
-//	{"summary": "...", ...}
-//	# confidence
-//	0.95
-//	# selected_workflow
-//	{"workflow_id": "...", ...}
-//	# alternative_workflows
-//	[{"workflow_id": "...", ...}]
-//
-// This function extracts each section's content, assembles an llmResponse,
-// and maps it to InvestigationResult via parseLLMFormat.
-func parseSectionHeaders(content string, logger logr.Logger) (*katypes.InvestigationResult, error) {
-	sections := extractSections(content)
-	if len(sections) == 0 {
-		return nil, fmt.Errorf("no section headers found")
-	}
-
-	assembled := make(map[string]json.RawMessage)
-
-	for header, body := range sections {
-		body = strings.TrimSpace(body)
-		if body == "" {
-			continue
-		}
-		assembled[header] = sectionBodyToJSON(body)
-	}
-
-	if len(assembled) == 0 {
-		return nil, fmt.Errorf("no parseable section content found")
-	}
-
-	compositeJSON, err := json.Marshal(assembled)
-	if err != nil {
-		return nil, fmt.Errorf("assembling section headers: %w", err)
-	}
-
-	return parseLLMFormat(string(compositeJSON), logger)
-}
-
-// sectionBodyToJSON converts a single trimmed, non-empty section body into
-// its JSON representation: arrays and embedded JSON objects are preserved
-// as-is, and bare scalars (numbers, "true"/"false"/"null", or unquoted
-// strings like "critical") are coerced into valid JSON.
-func sectionBodyToJSON(body string) json.RawMessage {
-	// Arrays (e.g. alternative_workflows) must be preserved as-is.
-	// extractJSON only finds objects, so it would strip the enclosing [].
-	if body[0] == '[' {
-		return json.RawMessage(body)
-	}
-
-	if jsonBody := extractJSON(body); jsonBody != "" {
-		return json.RawMessage(jsonBody)
-	}
-
-	// Scalar values: numbers ("0.95"), booleans ("true"/"false"), or strings ("critical")
-	quotedOrRaw := body
-	if quotedOrRaw != "true" && quotedOrRaw != "false" && quotedOrRaw != "null" {
-		if len(quotedOrRaw) > 0 && quotedOrRaw[0] != '"' {
-			if _, err := json.Number(quotedOrRaw).Float64(); err != nil {
-				quotedOrRaw = `"` + strings.ReplaceAll(quotedOrRaw, `"`, `\"`) + `"`
-			}
-		}
-	}
-	return json.RawMessage(quotedOrRaw)
-}
-
-// extractSections splits content on lines matching "# <header_name>" and returns
-// a map of header → body text. Recognizes headers with and without markdown fencing.
-func extractSections(content string) map[string]string {
-	knownHeaders := map[string]bool{
-		"root_cause_analysis":   true,
-		"selected_workflow":     true,
-		"alternative_workflows": true,
-		"confidence":            true,
-		"severity":              true,
-		"actionable":            true,
-		"investigation_outcome": true,
-		"detected_labels":       true,
-	}
-
-	sections := make(map[string]string)
-	lines := strings.Split(content, "\n")
-	var currentHeader string
-	var currentBody strings.Builder
-
-	flush := func() {
-		if currentHeader != "" {
-			sections[currentHeader] = currentBody.String()
-		}
-	}
-
-	for _, line := range lines {
-		trimmed := strings.TrimSpace(line)
-		if strings.HasPrefix(trimmed, "# ") {
-			candidate := strings.TrimSpace(trimmed[2:])
-			if knownHeaders[candidate] {
-				flush()
-				currentHeader = candidate
-				currentBody.Reset()
-				continue
-			}
-		}
-		if currentHeader != "" {
-			currentBody.WriteString(line)
-			currentBody.WriteString("\n")
-		}
-	}
-	flush()
-
-	return sections
-}
-
-// mergeNestedRemediationTarget checks for a nested root_cause_analysis.remediation_target
-// (or camelCase remediationTarget) when the flat parse path succeeded but
-// RemediationTarget is still empty. This handles hybrid JSON where the LLM
-// returns flat rca_summary/workflow_id alongside a nested RCA object.
-func mergeNestedRemediationTarget(result *katypes.InvestigationResult, jsonStr string) {
-	if result.RemediationTarget.Kind != "" {
-		return
-	}
-	var resp llmResponse
-	if err := json.Unmarshal([]byte(jsonStr), &resp); err != nil {
-		return
-	}
-	rca := resp.resolvedRCA()
-	if rca == nil {
-		return
-	}
-	if t := rca.resolvedTarget(); t != nil {
-		result.RemediationTarget = katypes.RemediationTarget{
-			Kind:       t.Kind,
-			Name:       t.Name,
-			Namespace:  t.Namespace,
-			APIVersion: t.APIVersion,
-		}
-	}
-}
-
-// mergeNestedInvestigationAnalysis extracts investigation_analysis from a nested
-// root_cause_analysis object when the flat parse path succeeded but the field
-// is empty. Handles hybrid JSON where the LLM returns flat rca_summary alongside
-// a nested RCA containing investigation_analysis (#724 F2).
-func mergeNestedInvestigationAnalysis(result *katypes.InvestigationResult, jsonStr string) {
-	if result.InvestigationAnalysis != "" {
-		return
-	}
-	var resp llmResponse
-	if err := json.Unmarshal([]byte(jsonStr), &resp); err != nil {
-		return
-	}
-	rca := resp.resolvedRCA()
-	if rca == nil {
-		return
-	}
-	result.InvestigationAnalysis = rca.InvestigationAnalysis
-}
-
-// mergeNestedSelectedWorkflow extracts selected_workflow fields from a nested
-// JSON object when the flat parse path succeeded but WorkflowID is empty.
-// Handles hybrid JSON where the LLM returns flat rca_summary alongside a nested
-// selected_workflow (#1431). Top-level workflow_id (already populated by
-// json.Unmarshal) takes precedence.
-func mergeNestedSelectedWorkflow(result *katypes.InvestigationResult, jsonStr string) {
-	if result.WorkflowID != "" {
-		return
-	}
-	var resp llmResponse
-	if err := json.Unmarshal([]byte(jsonStr), &resp); err != nil {
-		return
-	}
-	if resp.Workflow == nil {
-		return
-	}
-	result.WorkflowID = resp.Workflow.WorkflowID
-	result.ExecutionBundle = resp.Workflow.ExecutionBundle
-	if resp.Workflow.Confidence > 0 {
-		result.Confidence = resp.Workflow.Confidence
-	}
-	result.Reason = resp.Workflow.Rationale
-	result.WorkflowRationale = resp.Workflow.Rationale
-	result.ExecutionEngine = resp.Workflow.ExecutionEngine
-	if resp.Workflow.Parameters != nil {
-		if result.Parameters == nil {
-			result.Parameters = make(map[string]interface{})
-		}
-		for k, v := range resp.Workflow.Parameters {
-			result.Parameters[k] = v
-		}
-	}
-}
-
-// applyFlatFields applies LLM-provided flat fields to the result:
-// - actionable: sets IsActionable, synthesizes warning, applies confidence floor
-// - investigation_outcome: maps to outcome routing fields (HR is derived, not propagated)
-func applyFlatFields(result *katypes.InvestigationResult, flat flatLLMFields) {
-	if flat.Severity != "" && result.Severity == "" {
-		result.Severity = flat.Severity
-	}
-
-	if flat.Actionable != nil && !*flat.Actionable {
-		falseVal := false
-		result.IsActionable = &falseVal
-		// When investigation_outcome=problem_resolved, the outcome handler synthesizes
-		// its own warning ("Problem self-resolved"). Adding the generic "not actionable"
-		// warning would cause AA's response processor to route to NotActionable (Outcome D)
-		// instead of ProblemResolved (Outcome A). KA Python had the same precedence:
-		// the resolved outcome is authoritative over the actionable flag.
-		if flat.InvestigationOutcome != "problem_resolved" {
-			result.Warnings = append(result.Warnings, notActionableWarning)
-		}
-		if result.Confidence < confidenceFloor {
-			result.Confidence = confidenceFloor
-		}
-	}
-
-	if flat.InvestigationOutcome != "" {
-		result.InvestigationOutcome = flat.InvestigationOutcome
-		ApplyInvestigationOutcome(result, flat.InvestigationOutcome)
-	}
-
-	// #301: Contradiction override — when the outcome is problem_resolved but
-	// the parser-derived HR was set (e.g., via inconclusive outcome fallback),
-	// the resolution takes precedence. Defense-in-depth per HAPI parity.
-	if flat.InvestigationOutcome == "problem_resolved" && result.HumanReviewNeeded {
-		result.HumanReviewNeeded = false
-		result.HumanReviewReason = ""
-	}
-}
 
 // ApplyInvestigationOutcome maps KA-style investigation_outcome values
 // to is_actionable/needs_human_review/human_review_reason fields.
